03_etl/requirements.txt
asgiref==3.5.0
async-timeout==4.0.2
backoff==1.11.1
certifi==2021.10.8
Deprecated==1.2.13
Django==3.2
django-split-settings==1.1.0
elastic-transport==8.1.1
elasticsearch==7.7.0
Faker==13.2.0
flake8==4.0.0
gunicorn==20.1.0
mccabe==0.6.1
packaging==21.3
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
redis==4.2.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
wrapt==1.14.0
